requirements.txt
ftfy
accelerate>=1.2.1
einops
diffusers>=0.32.0
sentencepiece>=0.2.0
protobuf
torch
torchaudio
numpy<2.0
Pillow
imageio
imageio-ffmpeg
tqdm